Add specific data point GET method
All the planned GET methods are now implemented. This API is planned to
only allow safe operations on the database, so it is unlikely that
methods for modifying the database will be added in the future.

Also added validation of the inputs in the URIs to prevent SQL injection
attacks. The string that is passed through the URI is checked against
the list of colleges, tables, and data types in the database before
executing any SQL statement.