Update API to v2.0
Colleges are now specified and reported with a college_id value rather
than a name. This is because some colleges share the same name, making
the id the only unique identifier.

Many changes were made to the API calls. These changes were made to
allow retrieval of values of a specific data type over a year or range
of years. Further changes were made to make the retrieval of data type
and college data uniform, and to also allow the retrieval of global
(non-year-specific) data.

See the documentation for more details.