Make CSCExplorer compatible with IE11
Three changes have been made to allow CSCExplorer to work on IE11:

Criteria objects are no longer defined by the 'class' keyword. These
objects are now functions that contain data depending on the index
passed to the function.

Object.defineProperty is now used to create an object with a variable
property name rather than the previous implicit creation method which is
not compatible with IE.

Arrays.prototype.findIndex is no longer used to get the index of the
selected category for each criteria in the search, as this method does
not work with IE. A for loop that serves the same purpose has been
added.